Added counters for expected and actual exposure duration. The expected duration counter is used instead of the actual exposure duration counter in runs when we calculate the total number of interactions, as in these runs there are no infections.

+++ src/Types.h
@@ -25,7 +25,7 @@ using PartitionId = int;
 #define PARTITION_ID_PRINT_TYPE "%d"
 
 // Disease states
-using DiseaseState = int16_t;
+using DiseaseState = int;
 
 // Event types
 using EventType = char;
